% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagion-test.R
\name{ContagionTestST}
\alias{ContagionTestST}
\title{Shalizi and Thomas (2011) Contagion Test}
\usage{
ContagionTestST(df)
}
\arguments{
\item{df}{A dataframe of values where each column is a different year, and each row is a different node}
}
\value{
A dataframe of coefficients
}
\description{
Shalizi and Thomas (2011) Contagion Test
}
