% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{density_graph}
\alias{density_graph}
\title{Plots a density graph with the results from the split-halves test}
\usage{
density_graph(
  model,
  n,
  mean.xcoord = 0,
  mean.ycoord = 1,
  pval.xcoord = 0,
  pval.ycoord = 0,
  title = ""
)
}
\arguments{
\item{model}{A dataframe with three columns that contain the results of the split-halves test}

\item{n}{Number of iterations through split-halves test. Should be the same as the 'iterations' value input to lag_pc_test().}

\item{mean.xcoord}{The x-coordinate for where the mean value should be displayed on the graph. Default is 0.}

\item{mean.ycoord}{The y-coordinate for where the mean value should be displayed on the graph. Default is 1.}

\item{pval.xcoord}{The x-coordinate for where the p-value should be displayed on the graph. Default is 0.}

\item{pval.ycoord}{The y-coordinate for where the p-value should be displayed on the graph. Default is 0.}

\item{title}{The title to be displayed on the plot. Default is an empty string.}
}
\description{
Plots a density graph with the results from the split-halves test
}
