% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_pc_test.R
\name{lag_pc_test}
\alias{lag_pc_test}
\title{Run a parallelized network contagion test}
\usage{
lag_pc_test(
  df,
  iterations,
  cores = 1,
  difference = TRUE,
  threshold = 0.1,
  lagWin = 1,
  missingData = FALSE
)
}
\arguments{
\item{df}{A dataframe with a column of nodes, a column of years, and a column of values}

\item{iterations}{An integer denoting how many iterations of the test should be run}

\item{cores}{An integer denoting how many cores to use, default is 1}

\item{difference}{A binary indicator for whether df should be differenced. Default is TRUE.}

\item{threshold}{A threshold value for conducting a unit root test. If conducting a unit root test on df returns a p-value less than this value, the data will be differenced. Default is 0.1.}

\item{lagWin}{An integer indicating the lag window to be used. All the values within that window are averaged to get a single value. Default is 1.}

\item{missingData}{A binary indicator of whether 'df' includes missing data. If set to TRUE, ndiffs() will be used to test for stationarity.}
}
\value{
A dataframe with the coefficients from the contagion test
}
\description{
Run a parallelized network contagion test
}
\examples{
#running parallelized contagion tests on the data
ParallelContagionTest(df = polity90, iterations = 10000, cores = 3, difference = TRUE)

}
