% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals-calc.R
\name{PValsParallel}
\alias{PValsParallel}
\title{Parallelized p-value generation}
\usage{
PValsParallel(df, iterations, cores = 1)
}
\arguments{
\item{df}{A dataframe of values}

\item{iterations}{An integer denoting number of iterations}

\item{cores}{An integer denoting number of cores to run the parallelization on}
}
\value{
A numeric vector of p-values
}
\description{
Parallelized p-value generation
}
