% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-formatting.R
\name{STFormat}
\alias{STFormat}
\title{A function to transform data so it is in the proper format to be used with other functions in this package.}
\usage{
STFormat(df)
}
\arguments{
\item{df}{A dataframe where the first column contains the nodes (such as countries). The remaining columns each contain values for a single year.}
}
\value{
A dataframe with labels removed. Columns are unique years and rows are unique nodes.
}
\description{
A function to transform data so it is in the proper format to be used with other functions in this package.
}
\examples{
df <- cleaning(polityData)
}
