% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-contagion-test.R
\name{ParallelContagionTest}
\alias{ParallelContagionTest}
\title{Run a parallelized network contagion test}
\usage{
ParallelContagionTest(df, iterations, cores)
}
\arguments{
\item{df}{A dataframe with a column of nodes, a column of years, and a column of values}

\item{iterations}{An integer denoting how many iterations of the test should be run}

\item{cores}{An integer denoting how many cores to use}
}
\value{
A dataframe
}
\description{
Run a parallelized network contagion test
}
