% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvals-calc.R
\name{PValsParallel}
\alias{PValsParallel}
\title{Parallelized p-value generation}
\usage{
PValsParallel(df, iterations, cores = 1, difference = TRUE, threshold = 0.1)
}
\arguments{
\item{df}{A dataframe of time series values. Should be formatted such that each column contains values for a unique year. Each row is a different node.}

\item{iterations}{An integer denoting number of iterations to be run. Default is 100}

\item{cores}{An integer denoting number of cores to run the parallelization on. Default is 1.}

\item{difference}{A binary indicator for whether df should be differenced. Default is TRUE.}

\item{threshold}{A threshold value for conducting a unit root test. If conducting a unit root test on df returns a p-value less than this value, the data will be differenced. Default is o.1.}
}
\value{
A numeric vector of p-values, length is the same as the number of iterations
}
\description{
Parallelized p-value generation
}
\examples{
#using cleaning function to get data in the proper format
polity <- cleaning(polityData)

#200 iterations run on 2 cores
PValsParallel(df = polity, iterations = 200, cores = 2)

}
